+++ cluster-autoscaler/processors/provreq/provisioning_request_injector.go
@@ -42,7 +42,7 @@ var SupportedProvisioningClasses = []string{v1beta1.ProvisioningClassCheckCapaci
 
 // ProvisioningRequestPodsInjector creates in-memory pods from ProvisioningRequest and inject them to unscheduled pods list.
 type ProvisioningRequestPodsInjector struct {
-	client provisioningRequestClient
+	client provreqclient.ProvisioningRequestClient
 	clock  clock.PassiveClock
 }
