+++ cluster-autoscaler/provisioningrequest/provreqclient/client.go
@@ -41,15 +41,20 @@ const (
 	provisioningRequestClientCallTimeout = 4 * time.Second
 )
 
-// ProvisioningRequestClient represents client for v1beta1 ProvReq CRD.
-type ProvisioningRequestClient struct {
+type ProvisioningRequestClient interface {
+	ProvisioningRequests() ([]*provreqwrapper.ProvisioningRequest, error)
+	ProvisioningRequest(namespace, name string) (*provreqwrapper.ProvisioningRequest, error)
+}
+
+// provisioningRequestClient represents client for v1beta1 ProvReq CRD.
+type provisioningRequestClient struct {
 	client         versioned.Interface
 	provReqLister  listers.ProvisioningRequestLister
 	podTemplLister v1.PodTemplateLister
 }
 
 // NewProvisioningRequestClient configures and returns a provisioningRequestClient.
-func NewProvisioningRequestClient(kubeConfig *rest.Config) (*ProvisioningRequestClient, error) {
+func NewProvisioningRequestClient(kubeConfig *rest.Config) (*provisioningRequestClient, error) {
 	prClient, err := newPRClient(kubeConfig)
 	if err != nil {
 		return nil, fmt.Errorf("Failed to create Provisioning Request client: %v", err)
@@ -70,7 +75,7 @@ func NewProvisioningRequestClient(kubeConfig *rest.Config) (*ProvisioningRequest
 		return nil, err
 	}
 
-	return &ProvisioningRequestClient{
+	return &provisioningRequestClient{
 		client:         prClient,
 		provReqLister:  provReqLister,
 		podTemplLister: podTemplLister,
@@ -78,7 +83,7 @@ func NewProvisioningRequestClient(kubeConfig *rest.Config) (*ProvisioningRequest
 }
 
 // ProvisioningRequest gets a specific ProvisioningRequest CR.
-func (c *ProvisioningRequestClient) ProvisioningRequest(namespace, name string) (*provreqwrapper.ProvisioningRequest, error) {
+func (c *provisioningRequestClient) ProvisioningRequest(namespace, name string) (*provreqwrapper.ProvisioningRequest, error) {
 	v1Beta1PR, err := c.provReqLister.ProvisioningRequests(namespace).Get(name)
 	if err != nil {
 		return nil, err
@@ -91,7 +96,7 @@ func (c *ProvisioningRequestClient) ProvisioningRequest(namespace, name string) 
 }
 
 // ProvisioningRequests gets all ProvisioningRequest CRs.
-func (c *ProvisioningRequestClient) ProvisioningRequests() ([]*provreqwrapper.ProvisioningRequest, error) {
+func (c *provisioningRequestClient) ProvisioningRequests() ([]*provreqwrapper.ProvisioningRequest, error) {
 	v1Beta1PRs, err := c.provReqLister.List(labels.Everything())
 	if err != nil {
 		return nil, fmt.Errorf("error fetching provisioningRequests: %w", err)
@@ -108,7 +113,7 @@ func (c *ProvisioningRequestClient) ProvisioningRequests() ([]*provreqwrapper.Pr
 }
 
 // FetchPodTemplates fetches PodTemplates referenced by the Provisioning Request.
-func (c *ProvisioningRequestClient) FetchPodTemplates(pr *v1beta1.ProvisioningRequest) ([]*apiv1.PodTemplate, error) {
+func (c *provisioningRequestClient) FetchPodTemplates(pr *v1beta1.ProvisioningRequest) ([]*apiv1.PodTemplate, error) {
 	podTemplates := make([]*apiv1.PodTemplate, 0, len(pr.Spec.PodSets))
 	for _, podSpec := range pr.Spec.PodSets {
 		podTemplate, err := c.podTemplLister.PodTemplates(pr.Namespace).Get(podSpec.PodTemplateRef.Name)
